Demonstration of how to Backpack'ify this package.
In this Reddit thread
https://www.reddit.com/r/haskell/comments/28v6c9/backpack_an_mllike_module_system_for_haskell/cif36ed
@chrisdone mentioned tagstream-conduit as an example of a place where
lack of parametrized modules made it a bit annoying to write
entity parsing code that worked for both Text and ByteString.

In this commit, I give a refactoring of tagstream-conduit which
uses Backpack to the module encoding tricks and duplication.
The key component, entities/Entities.hs, is parametrized over the
type of String, Builder and Parser you want to use.  This
eliminates the duplication between Text/HTML/TagStream/ByteString.hs
and Text/HTML/TagStream/Text.hs, which are now just reexports of
appropriately instantiated versions of Entities.

There are three internal libraries:

    - 'types' defines common data types and "shim" modules for
      ByteString and Text.  (If I can convince text/bytestring
      to standardize naming for all their functions, I can
      reduce/eliminate these shim modules.)
    - 'entities' is the parametrized component, parametrized
      over the Str, Builder and Parser modules (the signatures
      required for these modules are in the hsig files.)
    - 'tagstream-conduit' is the main library and exposes the
      same outward facing interface as the old library.

Please take a look at the refactoring and let me know how you
like it; unfortunately, you'll need my development branches of
Cabal and GHC to actually build this package. In
particular, supposing that Backpack was supported on all
the versions of GHC/Cabal you wanted to target, would you
use this? What do you like/dislike about the system?

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>